samples/requirements.txt
featuremanagement
azure-appconfiguration-provider
azure-monitor-opentelemetry
azure-monitor-events-extension
quart
